requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile pyproject.toml
#
babel==2.14.0
    # via mkdocs-material
cairocffi==1.6.1
    # via cairosvg
cairosvg==2.7.1
    # via mkdocs-material
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cairocffi
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   mkdocs
    #   mkdocstrings
colorama==0.4.6
    # via mkdocs-material
cssselect2==0.7.0
    # via cairosvg
defusedxml==0.7.1
    # via cairosvg
faker==21.0.0
    # via pulp-docs (pyproject.toml)
ghp-import==2.1.0
    # via mkdocs
idna==3.6
    # via requests
jinja2==3.1.2
    # via
    #   mkdocs
    #   mkdocs-material
    #   mkdocstrings
    #   pulp-docs (pyproject.toml)
markdown==3.5.1
    # via
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markupsafe==2.1.3
    # via
    #   jinja2
    #   mkdocs
    #   mkdocstrings
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.5.3
    # via
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
mkdocs-autorefs==0.5.0
    # via mkdocstrings
mkdocs-material[imaging]==9.5.2
    # via pulp-docs (pyproject.toml)
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mkdocstrings==0.24.0
    # via pulp-docs (pyproject.toml)
packaging==23.2
    # via mkdocs
paginate==0.5.6
    # via mkdocs-material
pathspec==0.12.1
    # via mkdocs
pillow==9.5.0
    # via
    #   cairosvg
    #   mkdocs-material
platformdirs==4.1.0
    # via
    #   mkdocs
    #   mkdocstrings
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via mkdocs-material
pymdown-extensions==10.5
    # via
    #   mkdocs-material
    #   mkdocstrings
python-dateutil==2.8.2
    # via
    #   faker
    #   ghp-import
pyyaml==6.0.1
    # via
    #   mkdocs
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
regex==2023.10.3
    # via mkdocs-material
requests==2.31.0
    # via mkdocs-material
six==1.16.0
    # via python-dateutil
tinycss2==1.2.1
    # via
    #   cairosvg
    #   cssselect2
urllib3==2.1.0
    # via requests
watchdog==3.0.0
    # via mkdocs
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
